r/consensus: used chunked_vector in `raft::replicate`
When replicated batches in Redpanda we never use the capabilities of
record batch reader. It introduces additional burden as it needs to be
converted to vector of batches for batching and handling. Changed the
interface to accept `chunked_vector<model::record_batch>` or a single
batch instead of using readers.

Signed-off-by: Michał Maślanka <michal@redpanda.com>