raft: Make load_snapshot exception safe
The 'load_snapshot' method of the 'file_backed_stm_snapshot' class is
not exception safe. If the metadata loading operation throws the reader
is not closed properly which in turn triggers assertion.

This commit fixes this by handling exception and closing the reader
explicitly. If the metadata can't be loaded (snapshot is truncated) the
method rethrows the exception.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>