dt/client_quotas: use kafka quotas instead of configs
In addition to the straightforward mapping from configs to kafka quotas,
there are 2 test-behaviour changes:
 1. Previously the `ClusterRateQuotaTest` configured a default produce
and default fetch quota for all tests. This is removed and instead the
individual test methods set the same default produce or fetch quotas if
they are relevant to the test's behaviour. Eg.
`test_client_response_throttle_mechanism` only sets the fetch quota.
 2. In `test_client_quota_metrics` the metric for the partition mutation
quota is no longer incremented. This is because client quota metrics are
only recorded while there is at least 1 configured client quota.
Previously there was a client quota configured for the test class before
the cluster started (see (1)). This client quota was already in effect
at the time when the topic was created (by `RedpandaTest`), which meant
that the partition mutation metric got incremented. However, now the
quotas are configured after the topic is created, so the partition
mutation metric is not incremented.