bazel: use python toolchain for genrules
Without this one neat trick, genrules are using the system python3 to
run these python scripts (which are chmod +x python files with
 #!/usr/bin/env python3 at the top). This change explicitly uses the
Bazel toolchain. We must switch to the bazel toolchain because after
removing valgrind from the list of system packages in `install-deps.sh`
there is no longer python available in the docker image we build with.
It's better this way anyways.