redpanda: configurable sleep on crash loop
This introduces the node config `crash_loop_sleep`.

When redpanda detects that it reached the crash loop limit, instead of
terminating immediately, it should sleep for `crash_loop_sleep` seconds
before terminating.

This is useful especially in a Kubernetes environment where setting this
value allows customers to have ssh access into a crashlooping pod for a
short window of time.