bazel: change benchmarks to binaries
AFAIK this is standard practice in Bazel, we don't want to run these as
normal tests (we currently exclude them in CI), so just make them normal
binaries. We will need logic to schedule these, for example in CI we
don't run these with --overprovisioned, meaning they get pinned to a
core and if we run these in parallel we'll want to evenly distribute
these between cores. I'll leave the scheduling bit for another time, but
for now we just move the cc_bench rule to output a binary isntead of a
test.