iceberg: add method for building id->snapshot map
Callers that need to do non-trivial computation on the table metadata
shouldn't be tempted to use the existing vector in the metadata (left
simple to keep serialization simple).

This commit adds a method to build a map from the snapshots vector keyed
by the snapshot id to make it less tempting to write quadratic code.