serde: fix enum handling with c++23
The is_serde check enabled in c++23 (prior to this commit) checked only
for scoped enums. So the SFINAE magic for enum wasn't being triggered
for non-scoped enum. Instead, it was probably being handled by
scalar/integer encoding. Indeed, a not-enum guard is present on scalar
customization points, but, not-enum in c++23, again, means not scoped
enum.

Long ago we informally banned non-scoped enums. Turns out, there was a
non-scoped enum being used in serde. See previous commit for that.

This commit changes serde so that it retains the previous behavior that
worked in c++20, namely, the use of std::is_enum which checks the
superset of scoped and non-scoped enums.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>